requirement.txt
absl-py==1.0.0
asgiref==3.4.1
astunparse==1.6.3
backports.entry-points-selectable==1.1.1
CacheControl==0.12.10
cachetools==4.2.4
cachy==0.3.0
certifi==2021.10.8
charset-normalizer==2.0.9
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
distlib==0.3.4
Django==4.0
django-cors-headers==3.10.1
django-extensions==3.1.5
django-request-logging==0.7.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.4.0
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
html5lib==1.1
idna==3.3
importlib-metadata==4.10.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==21.8.0
libclang==12.0.0
lockfile==0.12.2
Markdown==3.3.6
msgpack==1.0.3
numpy==1.21.5
oauthlib==3.1.1
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==20.9
pandas==1.3.5
pastel==0.2.1
pexpect==4.8.0
Pillow==8.4.0
pkginfo==1.8.2
platformdirs==2.4.0
poetry==1.1.12
poetry-core==1.0.7
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pylev==1.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
shellingham==1.4.0
six==1.16.0
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tomlkit==0.8.0
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.10.0
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
